im: Split tests into separate modules
